rename film-view to FilmView
