+++ app/components/FilmView.tsx
